update 6 files and delete 14 files
