--- memorymarker/question_generator/highlight_to_question.py
@@ -1,20 +1,0 @@
-from typing import TYPE_CHECKING, Protocol
-
-if TYPE_CHECKING:
-    from iterpy.iter import Iter
-
-    from memorymarker.document_providers.contextualized_highlight import (
-        ContextualizedHighlight,
-    )
-    from memorymarker.question_generator.expanded_pipeline import PromptWithLineage
-
-
-class HighlightToQuestion(Protocol):
-    def __call__(
-        self, highlights: "Iter[ContextualizedHighlight]"
-    ) -> "Iter[PromptWithLineage]":
-        ...
-
-    @property
-    def name(self) -> str:
-        ...
