--- memorymarker/question_generator/question_generator_evaluator/type.py
@@ -1,3 +1,0 @@
-class SupportsIdentity:
-    def identity(self, pipeline_id: str, highlight: str) -> int:
-        return f"{pipeline_id}_{highlight}".__hash__()
