--- memorymarker/question_generator/question_generator_evaluator/types.py
@@ -1,1 +1,0 @@
-
